pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "AI Second Brain — Pydantic AI agents with Mem0 + Supabase backend."
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[anthropic]>=0.1.0",
    "mem0ai",
    "supabase",
    "python-dotenv",
    "pydantic-settings",
    "fastmcp",
    "click",
]

[project.optional-dependencies]
ollama = ["pydantic-ai[ollama]"]
dev = [
    "pydantic-ai[ollama]",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
brain = "second_brain.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
